feat. flip 적용(미완)
- test. player 카메라 셰이크 끔

+++ BeatHeart.h
@@ -10,7 +10,7 @@ private:
 	queue<unsigned int> beats; // 심장은 플레이어의 비트 성공/미스와 관계없이 정박에 뛰어야 함
 
 	float alpha;
-	int angle;
+	float angle;
 
 public:
 	BeatHeart();
